Split patching into it's own module to make file lengths nicer
